Using more robust `findPathJS` from resize-compute-js.
Path computation is not charged to the parallel critical path in the
statistics and the earlier code was having issues (extracting
undefined at semi-unpredictable times).

In short: Better to just unify the two code bases than spend time
debugging problems like that while the underlying PJS extensions to JS
engine are in an unstable state.