Various changes to try to get performance results to be competitive.
I took out some branching in an attempt to eliminate bailouts (mostly
when I was just guessing at what might be wrong).  But the crucial
thing was making the functions monomorphic by "cloning them" (here, by
the brute-force method of cut-and-paste and renaming the clones, and
then making each callsite call a distinct clone).

The necessity of doing this in the first place is quite disappointing,
but I think we are in discussions to try to improve matters, so
hopefully this sort of hack will not be necessary in the future.